KTIJ-13539 [kotlin] Move initialization of `KotlinFacetEditorGeneralTab` to `createComponent`
This will ensure that, when a single Kotlin facet is selected,
its editor is fully initialized and properly displayed

Also, remove initialization calls from the places where they
are not required. `MultipleKotlinFacetEditor` is created
when all the components for the sub-editors are already
created, and they should be already initialized. `apply`
call also should be invoked only when the component itself
has been already created

N.B. The fact that initialization of the editor
is no longer lazy should not cause freezes like
KTIJ-26280, because of the IDEA-327311 - the
multi-editor for all facets is no longer created
by default, so now only the explicitly selected
editors will actually be created and
initialized. Moreover, since all the editors
for the multi-selection editor are fully
created beforehand (including `createComponent`
call on each of them), this version of code
actually should not have any difference
from the old one in the sense of performance.

^KTIJ-13539 Fixed

(cherry picked from commit fdb123beabd9e6932fc0b41f5fb48dc9313f13d1)

IJ-CR-113859

GitOrigin-RevId: bdd9de3aa648d8dcfd8d3e244deb7974152695cd