K2: Update reference shortener to handle import alias
The existing reference shortener does not use import alias when it
shortens a symbol. Instead, it adds a new import directive for the
symbol that is already imported.

As we updated the analysis for the reference shortener:
 1. When shortening a symbol, check whether the symbol is already
    imported.
 2. If it is already imported by an import alias, keep the symbol
    reference expression and the import alias as a string together in
    `ShortenCommand`.

This commit updates PSIs based on the ShortenCommand.
`ShortenCommand.invokeShortening()` will do the following:

 1. If the expression to shorten has import alias,
    - Replace the target symbol with the import alias, and
    - Delete the qualifier from the target expression.
 2. Otherwise, just delete the qualifier from the target expression.
    (the same as its previous behavior)

For example, when we have the following code:
```
import my.package.foo as bar
.. my.package.foo(arg0, arg1, arg2) ..  // target symbol is foo
```

`ShortenCommand.invokeShortening()` will first update the call to
`my.package.bar(arg0, arg1, arg2)`, and then delete the qualifier to
shorten it to `bar(arg0, arg1, arg2)`.

KTIJ-27205