[Android] Remove approximate inline line number calculation.
It does not work very well and it is only used when debugging
on Android device before Android O. That means that it is only
used when debuggin on Android devices that are more than 6 years
old.

In addition, the code uses analysis in a non-cancellable read
action which leads to occasional hangs when used.